Fixes select whole note issue by avoiding side-effect
Since we were using a forEach we were always iterating through all selectors. This would've been
ok if it wasn't for the side-effect within that forEach. Selecting a note is an action which
modifies the markers. We keep the reference to the "old" markers and pass those into
isSelectionEntirelyWithinNote which causes things to break.

Will have a think about how we can make the noting APIs less foot gunny. If we can avoid side
effects more that should reduce similar  errors.